requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
boto3==1.26.157
botocore==1.29.157
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-daraja==1.3.0
django-environ==0.10.0
django-shortcuts==1.6
django-storages==1.13.2
django-taggit==3.0.0
emoji==2.2.0
environ==1.0
executing==1.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
lxml==4.9.2
matplotlib-inline==0.1.6
msal==1.22.0
mypy==1.3.0
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.2
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
